# openapi.yaml
openapi: 3.0.0
paths:
...
  /users/email:
    post:
      operationId: UserControllerV2_createWithEmail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserWithEmailDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Sign up with email (no password yet)
      tags: &a1
        - Users
  /users/google:
    post:
      operationId: UserControllerV2_createWithGoogle
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserWithGoogleDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Sign up / continue with Google
      tags: *a1
  /users/link-google:
    post:
      operationId: UserControllerV2_linkGoogleAccount
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkGoogleAccountDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Link a Google account to an existing user
      tags: *a1
  /users/password:
    post:
      operationId: UserControllerV2_createPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePasswordDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Set a password for a user that has none
      tags: *a1
    put:
      operationId: UserControllerV2_updatePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Change an existing password
      tags: *a1
  /users/{id}/has-password:
    get:
      operationId: UserControllerV2_hasPassword
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HasPasswordResponse"
      summary: Check whether a user has a password set
      tags: *a1
  /users/{id}:
    get:
      operationId: UserControllerV2_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Get a user by id
      tags: *a1
    delete:
      operationId: UserControllerV2_deleteAccount
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Soft-delete the account
      tags: *a1
  /users/{id}/profile:
    patch:
      operationId: UserControllerV2_updateProfile
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Bulk update profile fields
      tags: *a1
  /users/{id}/first-name:
    patch:
      operationId: UserControllerV2_updateFirstName
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFirstNameDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Update first name
      tags: *a1
  /users/{id}/last-name:
    patch:
      operationId: UserControllerV2_updateLastName
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLastNameDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Update last name (null to clear)
      tags: *a1
  /users/{id}/email:
    post:
      operationId: UserControllerV2_requestEmailChange
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestEmailChangeDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
      summary: Request an email change (sends a code to the new email)
      tags: *a1
  /users/{id}/email/confirm:
    post:
      operationId: UserControllerV2_confirmEmailChange
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmEmailChangeDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Confirm an email change with the emailed code
      tags: *a1
  /users/{id}/avatar:
    patch:
      operationId: UserControllerV2_updateAvatar
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Update avatar (url + key from file upload)
      tags: *a1
  /users/{id}/business:
    patch:
      operationId: UserControllerV2_setBusinessId
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetBusinessIdDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
      summary: Attach a business to the user
      tags: *a1
...
  /admin/users/{id}:
    get:
      operationId: AdminManageUsersController_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Get a single user by id
      tags: *a4
    delete:
      operationId: AdminManageUsersController_softDeleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Soft-delete a user (sets deleted=true)
      tags: *a4
...
info:
  title: Ventura API
  description: Ventura backend API.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: JWT access token from /auth sign-in.
  schemas:
    CreateUserWithEmailDto:
      type: object
      properties:
        firstName:
          type: string
          example: Ada
        email:
          type: string
          example: ada@example.com
          format: email
      required:
        - firstName
        - email
    UserResponse:
      type: object
      properties:
        id:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
        shortId:
          type: string
          example: aB3xY9kP
        firstName:
          type: string
          example: Ada
        lastName:
          type: object
          nullable: true
          example: Lovelace
        email:
          type: string
          example: ada@example.com
          format: email
        googleId:
          type: object
          nullable: true
        avatarUrl:
          type: object
          nullable: true
        avatarKey:
          type: object
          nullable: true
        businessId:
          type: object
          nullable: true
        isSystem:
          type: boolean
          example: false
        isActive:
          type: boolean
          example: true
        isEmailVerified:
          type: boolean
          example: false
        deleted:
          type: boolean
          example: false
        deletedAt:
          type: string
          nullable: true
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - shortId
        - firstName
        - email
        - isSystem
        - isActive
        - isEmailVerified
        - deleted
        - createdAt
        - updatedAt
    CreateUserWithGoogleDto:
      type: object
      properties:
        firstName:
          type: string
          example: Ada
        lastName:
          type: string
          example: Lovelace
        email:
          type: string
          example: ada@example.com
          format: email
        googleId:
          type: string
          example: google-oauth2|1234567890
        avatarUrl:
          type: string
          example: https://example.com/a.png
      required:
        - firstName
        - email
        - googleId
    GoogleLinkPreferencesDto:
      type: object
      properties:
        firstName:
          type: string
          enum:
            - keep
            - update
        lastName:
          type: string
          enum:
            - keep
            - update
        avatarUrl:
          type: string
          enum:
            - keep
            - update
    LinkGoogleAccountDto:
      type: object
      properties:
        email:
          type: string
          format: email
        googleId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        preferences:
          $ref: "#/components/schemas/GoogleLinkPreferencesDto"
      required:
        - email
        - googleId
    CreatePasswordDto:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        newPassword:
          type: string
          minLength: 12
          example: sup3r-Secret!pw
      required:
        - userId
        - email
        - newPassword
    UpdatePasswordDto:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 12
          example: new-Passw0rd!yy
      required:
        - userId
        - email
        - oldPassword
        - newPassword
    HasPasswordResponse:
      type: object
      properties:
        hasPassword:
          type: boolean
          example: true
      required:
        - hasPassword
    UpdateProfileDto:
      type: object
      properties:
        firstName:
          type: string
          example: Ada
        lastName:
          type: object
          nullable: true
          example: Lovelace
        avatarUrl:
          type: object
          nullable: true
    UpdateFirstNameDto:
      type: object
      properties:
        firstName:
          type: string
          example: Ada
      required:
        - firstName
    UpdateLastNameDto:
      type: object
      properties:
        lastName:
          type: object
          nullable: true
          example: Lovelace
      required:
        - lastName
    RequestEmailChangeDto:
      type: object
      properties:
        newEmail:
          type: string
          example: new-address@example.com
          format: email
      required:
        - newEmail
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Logged out.
      required:
        - message
    ConfirmEmailChangeDto:
      type: object
      properties:
        code:
          type: string
          example: "123456"
          minLength: 6
          maxLength: 6
      required:
        - code
    UpdateAvatarDto:
      type: object
      properties:
        avatarUrl:
          type: object
          nullable: true
          example: https://example.com/a.png
        avatarKey:
          type: object
          nullable: true
          example: avatars/abc.png
      required:
        - avatarUrl
        - avatarKey
    SetBusinessIdDto:
      type: object
      properties:
        businessId:
          type: string
          example: 665f1b2c3d4e5f6a7b8c9d0e
      required:
        - businessId
...